Update patch set 14
Patch Set 14:

> that means the bundle was compiled with compiler source/target level 9 although its BREE says: Bundle-RequiredExecutionEnvironment: JavaSE-1.6

This is overridden in the pom.xml by target-platform-configuration/executionEnvironment setting JavaSE-9.
JavaSE-9 was necessary to fix the initial error which is that the product couldn't be provisioned because of missing a.jre.javase 9.0.0 IU in the target-platform.

But changing the EE in target-platform-configuration seems to also change the compilation target level (to 9 in that case), so the workaround here is to override the compilation target to something lower.

I can also try the following workarounds if you prefer (not sure they'd work):
* move test bundle to JavaSE-9 as BREE so the newer a.jre.javase unit is in the TP and surefire can resolve it
* force inclusion of a.jre.javase 9.0.0 as requirement in TP, but I'm afraid some other layer in Tycho would remove it.

Patch-set: 14