Update patch set 14
Patch Set 14:

>JavaSE-9 was necessary to fix the initial error which is that the product 
>couldn't be provisioned because of missing a.jre.javase 9.0.0 IU in the target-
> platform.

is this a general issue? are products now always requiring a.jre.javase 9.0.0 and builds against old p2 repos not providing this IU will fail?

what's the relation to https://bugs.eclipse.org/bugs/show_bug.cgi?id=387701 , shouldn't this remove the dependency from products to a.jre.javase?

Will people have to do the same workaround in their builds? If so, this would be a breaking change and I think it would be hard to argue that everybody should do a workaround starting with the new Tycho version...

Patch-set: 14