Update patch set 14
Patch Set 14: Code-Review-1

> or maybe this only happens for the special case of surefire "p2 installed product" mode?

Yes, so far it only happens in that case.
But we've identified an issue when using the proposed snapshot to build linuxtools. I'm investigating and will hopefully come up with a new patch that works for linuxtools, and maybe even make the workaround useless.

Patch-set: 14
Label: Code-Review=-1