Update patch set 16
Patch Set 16: Code-Review+2

The big issue with the test here is that the EE of the test is inferior to what's required by product, and when resolving dependencies, Tycho decides remove the newer "a.jre.javase" from the TP, and then from the provisioning operation of the test.

The simplest remediation, if we want this test to succeed, is to increase the BREE to JavaSE-9 so the test provides the capability (and request transitively the right a.jre.javase unit) to succeed.
This also requires to configure the tycho-surefire-plugin of this specific test project to set <useJDK>BREE</useJDK> and JavaSE-9 configured as toolchain to avoid facing the incompatible class version issue.
Does this seem proper enough to you? I can submit a newer version of this patch trying it.

But that doesn't solve the actual issue beneath it which is that in the case of product provisioning by tycho-surefire-plugin when test doesn't use a BREE which is the one built-in with product. This is a story for the other bug https://bugs.eclipse.org/bugs/show_bug.cgi?id=387701.

Patch-set: 16