Bug 532775 - Always include latest EE when resolving dependencies
Transitive dependencies may require a more recent EE than the one
of the bundle being built. With EE now considered as requirements/
capabilities by p2, this situation will resolve in (frequent)
resolution error with need for clients to adapt their build config.

To workaround this problem, this adds the most recent know EE into
the resolution context, but to minimize the interference, we remove
the packages from tempoarary EE units if they're not the ones
requested by the module being built.

Change-Id: Id908d3a44c175ba2c1914d83237bc096f4efc209
Signed-off-by: Mickael Istria <mistria@redhat.com>