Update patch set 2
Patch Set 2:

Different tests do fail because the behavior we're implementing is explicitly what the tests are meant to prevent.
What would be interesting would be to 1st try a dependency resolution without adding the extra-units for latest osgi.ee, and if it fails, try again with additional unit.
I'm afraid that'd be complicated to implement well and expensive at usage.
@Jan: WDYT?

Patch-set: 2
Label: Verified=0