Update patch set 4
Patch Set 4:

yes I think we we need to keep adding both jre IUs 1.6.0 and 1.9.0 more or less forever as there are existing published p2 repos out there that have the hard dependency from products to a.jre IUs 1.6.0 and 9.0.0 and we want to be able to build against these repos. 
LAST_KNOW_EE should always move to latest known EE independently of that.

Patch-set: 4