Update patch set 1
Patch Set 1:

> What's the actual issue that triggered this patch?
 > In general, the default EE is not so important for the build
 > results so I'm curious about how things fail.

The test case I added is a simplified example, but in general I am encountering bundles that do not specify a EE, either in the pom or the manifest. Mostly these are documentation bundles, or otherwise do not contain any bytecode. In such a case the default EE is used and the resulting built artifact cannot be used on Eclipse on Java 1.8

Patch-set: 1