Update patch set 1
Patch Set 1: Code-Review-1

That's the issue. The EE used at build time should never be expected to be the EE that's required to run the bundle. It's an EE that's computed "as best" to guarantee build and test works. Depending on the dependency chain that was resolved, it can be an higher one than the minimal one.
I think it's more a matter of fixing the piece of login in the RPM build which compute the minimal EE to read the actual BREE in the bundle rather than assuming the EE resolved by Tycho in the minimal one.
I don't think Tycho should be changed for that.

Patch-set: 1
Reviewer: Gerrit User 587 <587@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1