Update patch set 1
Patch Set 1: Code-Review-1

This is likely the bare minimum that we would like to have, so that Gerrit jobs also dump stack traces when nearing the build timeout.

The code is mostly copied from EclipseTestRunner, so I guess it should work. I'll try to build and run with the change to validate, this will likely take a while as well (I am new to Maven and Tycho).

Patch-set: 1
Label: Code-Review=-1