Update patch set 2
Patch Set 2:

I've moved the code to create dumps to e4.workbench3, see reviews:

https://git.eclipse.org/r/#/c/133831/ (moved code)
https://git.eclipse.org/r/#/c/133835/ (clean-up old code)

The code is now fairly easy to access. As far as re-using it goes, we can possibly add extra methods to TestableObject, to add the timeout timer. Since we need to access the timeout command line argument, I don't think we can add the dumps without any new methods to TestableObject.

I.e. test applications will need extra calls to activate the dump-on-timeout. I don't have a preference whether the extra calls are to TestableObject or to some other class.

Mickael, is this what you wished or have I misunderstood you? If this is what you pictured, we can review and merge the other changes. Then I'll update this change to use the moved code.

Patch-set: 2
Label: Verified=0