Update patch set 20
Patch Set 20:

We can try but I don't know how to do it. And I am sure I have no particular right in tycho to do it. I am not commiter. So If you can do it and help me about this, it will be wonderfull ? In this case, you ust remove the target and change the junit test to use the target, instead of the localy defined.

Patch-set: 20
Label: Verified=0