Update patch set 5
Patch Set 5:

(1 comment)

Jan, Mickael, I have no preference for using reflection or copying the code from EclipseTestRunner (which was now moved and should therefore be usable with reflection). Either is bad but seemingly necessary.

This change copies the code (as patch set 1 did). Let me know if you prefer reflection or there are other things to be changed in the new code.

Patch-set: 5