Update patch set 4
Patch Set 4:

forget my comment about TestMojo. In fact also compilation can fail with JDK 11. would be nice to have an integration test for both the compilation and the test failure.

If we change it as you propose in fact we change the BREE of all bundles to the current JDK if JDK >=11 is used. this should be announced.

also, need to clarify what happens in case useJDK=BREE for compilation and test runtime

Patch-set: 4
Label: Verified=0