Update patch set 9
Patch Set 9:

> I wonder why we do a second resolution instead of doing it right the first time directly in DefaultTychoResolver.resolveProject() , i.e. add our special handling for JDK >=11 there?

I think we might want to catch issues that happen with the BREE as well. Instances can be that older BREE doesn't provide a new API for which there is a bundle available on Orbit, and this bundle is missing in TP, we'd still like the build to fail saying it can't resolve the package with Java 8 even if running on Java 11.
Unless I misunderstood your comment, we'd loose these checks.

Patch-set: 9
Label: Verified=0