Update patch set 10
Patch Set 10:

> > (2 comments)
 > >
 > > see inline.
 > >
 > > I wonder why we do a second resolution instead of doing it right
 > > the first time directly in DefaultTychoResolver.resolveProject()
 > ,
 > > i.e. add our special handling for JDK >=11 there?
 > 
 > Let me try that. First answer is "because I'm still finding my path
 > in this area".

So I tried it and the issue found is that dependencyComputer.computeDependencies must happen within a block where the current JRE EE is used otherwise it will not generate classpath entries due to finding them in project EE. 
The short version is that if we want it to happen only once the logic which EE to use must be duplicated in two places DefaultTychoProject.resolveProject and DependencyComputer.computeDependencies. Jan, do you think I should go this path?

Patch-set: 10
Label: Verified=0