Update patch set 12
Patch Set 12: Code-Review+2

great work Alexander. Thanks for being persistent.

To be honest for this part of the codebase I have to reverse-engineer it myself.

Since the JDK started removing packages it's tough to satisfy both use cases:

a) correctly resolving packages removed from the JDK from bundles in the target platform instead by always using the current EE for resolution
b) failing the build if newer packages from the current JDK are used but the bundle claims to be able to run on older JDK according to its BREE

using a) by default if JDK >=11 strikes a good balance for usability I think.

Just in case you want to be extra strict at build time already, I think that the useJDK=BREE  option should still give you an error if you use packages not yet present in an older JDK (in this case the bundle BREE's JDK is used for the compilation classpath and compilation will fail if a newer package was used - to be double-checked if this holds true)

we should announce the new behaviour as soon as someone is using JDK >=11 for the build in the release notes.

Patch-set: 12
Label: Code-Review=+2