Update patch set 2
Patch Set 2:

(1 comment)

I'm not much aware of Maven parallel build and the meaning of threadSafe here. Does it mean that the mojo can run in parallel of any other mojo, or that it can run in parallel of itself only.
If the former, then synchronizing on a static mojo specific lock isn't enough.

Patch-set: 2
Label: Code-Review=0