Update patch set 3
Patch Set 3:

> The root causes are much deeper IMO: in p2 that should rebuild the available EE at startup according to the actually available ones at installation-time instead of assuming the EE are published statically and that a statically published EE is immediately available;

I agree, I think at one time p2 had a goal to provision another runtime other than the runtime p2 is executing in.  And that target runtime may be launched with a different JRE.  But I must admit that I have no idea how that would have worked in practice.  The framework implementation did away with statically defined Java profiles long ago and I think most others should do the same.

>  and in OSGi itself that the BREE are becoming meaningless and not practical now that Java has abandonned backward compatibility (removed APIs).

This is partially true, The BREE (which is deprecated) or the osgi.ee requirement namespace is more about defining the minimum Java level required to interpret the bytecodes of the classes in a bundle.  Now with OSGi R7 bundles are able to specify import package on the java.* packages also, so if they need specific sub packages of java.* that are outside of the java.base module they import-package header to specify requirements on them.  Any profiles that represent the running instance should also include the java.* packages that are available.

As to the statement on backwards compatibility.  I don't have specific knowledge of this, but my general impression is that after the initial work of cleaning up the mess that was JavaSE into a modular runtime, going forward they generally want to avoid breaking changes.

Patch-set: 3
Reviewer: Gerrit User 1197 <1197@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0