Update patch set 2
Patch Set 2:

> (1 comment)

(1) ECJ would reject -extdirs and -endorseddirs as well with source level 9 and above.
(2) As far as ECJ is concerned, when the source alone is passed, the target is set to same as source. But not vice versa. The compiler rejects when only target is passed. So, ignoring that scenario, like we are doing now, looks like a safe option to me.

Patch-set: 2
Reviewer: Gerrit User 663 <663@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0