Bug 551292 - Use the same plexus container as Maven 3.1.0+
Tycho uses plexus-container-default which doesn't make much sense as it's the container used by 
Maven 2.x while Tycho's minimal requirement is Maven 3.0.5. 
Maven 3.0.x uses sisu-inject-plexus but from the range of supported versions only Maven 3.0.5
uses it as next version is 3.1.0 which already uses org.eclipse.sisu.plexus.
Additional benefit of switching to org.eclipse.sisu.plexus is that as an eclipse project 
artifact it doesn't require CQ. The only downside (if we can call it that way ) is that 
version 3.0.5 will no longer be supported which should be fine considering that 
Maven 3.1.0 is released on 2013-07-15.

Change-Id: I76121d6e7212fab3c742d0cdc468a446dee2734e
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>