Update patch set 4
Patch Set 4: Code-Review+2

I think this could be a bit of a risky change as it changes the default behavior and some users may rely on the current behavior and dislike the new one which is more "tolerant".
So ideally, we should have some flag to control it, make the tolerant behavior opt-in...
But realistically, I think this change is for the better in the very vast majority of cases, so I'm willing to take the risk of maybe annoying a few users, and let those who are annoyed by it come and contribute remediation that fit their cases.

Patch-set: 4
Label: Code-Review=+2