Bug 551974 - [performance] DefaultBundleReader manifest cache too small
Use an unlimited manifest cache to avoid repeated disk access. Before
the change resolving the target platform of a custom eclipse application
took several minutes, spending the majority of the time in reading
manifests from .jar files again and again.

With the unlimited cache in place, the time spent for reading the
manifests from the jars drops from several minutes to 3 seconds. The
manifest cache contains 761 entries and uses 3.6 MB in my case.

Change-Id: I34175635754b57a6cfa002a595b6716209e96c69
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>