Bug 552066 - Respect src.includes/src.excludes for source features
When building the "base" feature, the bin.includes and bin.excludes
are used to populate the feature with resources based on feature
developers preference.
When building the "source" feature for a "base" feature, the
src.includes and src.excludes properties should be used to populate the
"source" feature with resources based ont he featue developers
prefernce.

The functions in this changeset are copied from PackageFeatureMojo.java
and AbstractTychoPackagingMojo.java and then updated to use the src
properties rather than the bin properties.

Change-Id: I5dcf57120b559d2bb63ceac04a6c33a03b38d199
Signed-off-by: Torbjörn Svensson <azoff@svenskalinuxforeningen.se>