Fix working sets in Tycho workspace
My previous change of the working set declarations was missing AND nodes
as top level for both the Tycho and TychoExtras working set. I've added
these now and tested with a copy of the working set definition in my
workspace. This now correctly separates the bundles in _either_ the
first or second working set.

Change-Id: Id05902ffb328830dfb89118df490a83ed996dabd
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>