Update patch set 1
Patch Set 1: Code-Review+1

I was just going to ask why you didn't upgrade all the other maven dependencies as well, when I noticed that the related bug is linked to an IPZilla one. So that answers the question already.

Can we upgrade at least scope=test dependencies like hamcrest without a CQ? I'm not so experienced in that area yet.

Patch-set: 1
Label: Code-Review=+1