Stop using downloaded Eclipse 3.6 to run integration tests
Using ancient eclipse version and in this fragile way of coping and
extracting makes it really hard for people to enhance the tests.
The requirement to have the whole eclipse zip and deltapack (not
produced for years) published to maven central complicated it further.
This was all found when trying to understand why ITs don't run with Java
11 and it happens that the equinox used in them can't run on modular
JVM.

Change-Id: I51f7a6ecb9a8ef7987eb8c1101336a24c82e90ad
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>