Bug 546360 - Some Tycho tests fail when building with Java 11
There is the issue that it was possible to resolve newer Java versions
than the running ones relying on the EE profile file. In order to keep
this possibility (although it will only work if no new APIs are actually
used) if the EE profile specified is newer than the running version it
fallbacks to the old profile file way.

Change-Id: I3c2a8af75571131a184e139ac0603f7c514b0d83
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>