Update patch set 1
Patch Set 1:

Mickael, I am afraid not, as this probably goes into surefire-booter/OSGi-classloader territory I am not familiar with. The failing IT was just meant as convenient steps to reproduce for you guys.

For what it's worth, it *may* be that the tycho-surefire-plugin just needs some special configuration, rather than a fix in code, but if so, I haven't found it yet.

Patch-set: 1