Update patch set 3
Patch Set 3: Code-Review-1

Can you explain why you needed to modify the tests for them to pass? Usually, having to modify the tests means that the change is not backward compatible, and we should avoid that (unless the behavior was a bug or is deprecated).
I'm putting a temporary -1 to track discussion is in progress, that doesn't necessarily mean that more change is needed. We'll see what needs to be done or not as we discuss.

Patch-set: 3