Update patch set 3
Patch Set 3:

Hello, I hesitated to make these changes to test, but default value for showWarning is supposed to be false, so having to set it as true to actually have warning displayed is supposed to be done,  in my opinion these tests should not have passed. 
Of course it shows a possible change of behavior, so even if it makes things work as they should, I understand you would prefer not to change existing behavior by default.
It is possible to have a fix without changing tests, by changing showWarning default value depending of the way other parameters are explicitely set, but it would make things a bit harder to understand I guess.

Patch-set: 3