Update patch set 3
Patch Set 3:

OK, I think that makes sense to change this as it's relatively low impact.
However, can we make it so that if `maven.compiler.failOnWarning` is set to true, warning are shown? Failing compilation when there are warnings and not showing them seems like a bad reporting.

Patch-set: 3