Update patch set 2
Patch Set 2: Verified+1 Code-Review+2

> @Alex: is it ok to merge now so it's available in 2.0.0-SNAPSHOT
 > soon?

OK. But this should be the last profile we add to Tycho. Equinox doesn't ship them anymore for a reason. In the future support for new Java versions should rely on bytecode version|compiler --release tag. And of course toolchain definition of a JVM of the given version so packages are known.

Patch-set: 2
Label: Verified=+1
Label: Code-Review=+2