Update patch set 2
Patch Set 2:

In fact I think none of them :-)

The main problem with that approach is: You want to enable multithreaded, but synchronize on the method block. That means that effectively this code can't be run in parallel.

And even though this might be assumed suitable, there is no grantee that the class is really threadsafe... it accesses some external services that might or might not be save to be called concurrently (even though it is synchronized here does not mean it is synchronized elsewhere).

So the first (and probably most timeconsuming) part of the solution must be to check each and every service used here if it is safe to call concurrently. If this is the case, and we assume that injection is only happening at constructiontime, you even don't need any synchronization at all if the method only operates on local and (semi) final fields.

Patch-set: 2
Reviewer: Gerrit User 171685 <171685@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0