#553434 Use one Maven Import per root folder
When importing all projects within one "Maven Import"-Task, an
infinitive loop occurred.
Now there are "Maven Import"-Tasks for each root folder and the "Locate
Nested Projects" flag is only set to true if there are nested projects.
With this approach, the infinitive loop does not occur.
For the user of the Oomph setup there is no difference.
The only drawback is that a "Maven Import"-Task has to be added in the
setup file if new root folders/projects get introduced.

Bug: 553434
Change-Id: I4e925dd45d91931d4a63ac769c57b9d0304a23e9
Signed-off-by: Martin Schreiber <MARTIN.schreiber@bachmann.info>