#553434 Disable auto building during setup
With the auto-build option set to false during the setup process, the
projects do get build the first time after all the projects got imported
and after the target platform was set.
This fixes most of the issues with build errors. There is still a
manuell step needed to call "Maven"->"Update Project..." for some
projects.

Bug: 553434
Change-Id: I14647024b323ce0eb16d291d5d641e99f9c0fc8f
Signed-off-by: Martin Schreiber <MARTIN.schreiber@bachmann.info>