Bug 563930 - Resolve system packages from toolchains when not in profile
When profile doesn't define system packages (case of Java 9+ with module Java),
Tycho tries to resolve the package from the defined toolchains when available
(using a companion jar to run against the toolchain to dynamically list packages),
and fails back to the current runtime system packages if no matching toolchain is found.

This requires propagating the mavenSession, toolchainManager and Logger from the mojos
to StandardExecutionEnvironment.

Change-Id: Ie3133c1867e50d22ac7f341b70cc03ed00d3e019
Signed-off-by: Mickael Istria <mistria@redhat.com>