Update patch set 4
Patch Set 4:

(10 comments)

Thanks for taking a look at this, this is btw related to the change https://git.eclipse.org/r/#/c/165028/

I know its a little bit bigger change but in the end far less complex as I first have feared about :-)

And it is a little bit rough atm as I have to deal with lots of "legacy" behavior and have to decide what should be changed now and what should be postponed to make the change not even bigger.

I tried to answer your comments and give some insights about the decisions that lead to the code as it is right now, so let me know if you think we can proceed for now with some minor adjustments or do you think the story should be enhanced right now with some more refactoring?

Patch-set: 4