Update patch set 4
Patch Set 4:

(10 comments)

I've put some more comments inline, mostly about naming, trying to better separate things, and making things less mutable.
As mentioned earlier, I think the general design is very good and I'm confident we'll be able to merge that in a near future; however, as it's critical pieces of code and existing APIs (or new ones), I think it's worth spending the effort on the review before we merge -and release- and intermediary step with APIs that we'll want to change again soon after.

Patch-set: 4