Update patch set 3
Patch Set 3:

> Thank you Jay!
 > @Alexander: so the only missing thing is to actually run the test.
 > Is there any particular thing that is blocking you on this topic,
 > or will you be able to provide that relatively soon?

My problem is I do not really have a starting point for how to setup the test, and how to make sure that it actually tests the right thing, especially with the toolchains.xml etc. Can you point me to a similar test?

Otherwise I fear this is out of my reach.

Patch-set: 3
Label: Verified=0