Update patch set 5
Patch Set 5:

The build failure has revealed an interesting detail here:

One of the test-repositories only contains a meta but no artifact repository what is uncommon but valid I think. Currently tycho fails in such a case, I'd like to change this so it only emits a warning if this is okay.

Patch-set: 5