Update patch set 5
Patch Set 5:

> Currently tycho fails in such a case

You mean currently with this patch; not with master?

> I'd like to change this so it only emits a warning if this is okay.

I don't think it's worth warning, as it's a totally valid cases for "logical" ius, or for categorization only. Actually feature could be replaced by such logical ius as well.
Moreover, it's not related to the current change.
So IMO, it's just better to keep everything as it.

Patch-set: 5