Use propagated ExecutionEnvironment for resolution
Instead of recomputing new ones that may be missing or have glitches, we
reuse the ones made available in the eeConfiguration (computed with
profiles and other settings).
This better separate concerns, remove some duplication, improve
factorization.

Some irrelevant EE tests have been removed as they should be moved to
tycho-core (with actual EE registry handling) instead of resolver.
Some tests were removed for cases of product hardcoding dependency on
`a.jre` as this case is deprecated and not supported any more.

Change-Id: I37244f1d789c483f439ebebfd096b0d972d55b3e
Signed-off-by: Mickael Istria <mistria@redhat.com>