Update patch set 2
Patch Set 2:

The prefix is used in compliance to how it works for pomless, so if people use pomless but later decide to have a pom.xml for advanced configuration there is no need to adjust the build.properties and vice versa.

Beside this, as we do not have control over the keys in build.properties it prevents name-clash with future PDE enhancements.

Patch-set: 2