Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Is the the test failure something you can reproduce locally?
> 
> I can, when running mvn verify locally.
> 
> But I cannot from Eclipse - whole test suite is green.
> 
> The stack trace does not really give any hint at all -- almost as if something gets swllowed by JUnit somewhere.
> 
> Using -X -debug outputs nothing helpful anywhere either :/

I did some more digging and tried to reproduce what goes on with that test method.

Using javac locally:

  367431_frameworkExtensionCompileAccessRules\bundle\src> javac tycho\testsun\TestClass.java
  tycho\testsun\TestClass.java:7: error: package com.sun.net.ssl.internal.ssl is not visible
          Security.addProvider(new com.sun.net.ssl.internal.ssl.Provider());
                                                         ^
    (package com.sun.net.ssl.internal.ssl is declared in module java.base, which does not export it)
  1 error

This is the expected situation that this test which it then solves by adding a framework extension bundle.

But now, when also specifying --release:

  367431_frameworkExtensionCompileAccessRules\bundle\src> javac --release tycho\testsun\TestClass.java
  tycho\testsun\TestClass.java:7: error: package com.sun.net.ssl.internal.ssl does not exist
          Security.addProvider(new com.sun.net.ssl.internal.ssl.Provider());

When --release not used, the default jrt-fs.jar of java 11 is used (which luckily still contains that class in 11, in 14 it's already gone), when --release is used the CtSym kicks in. But apparently the CtSym does not contain all internal sun classes. Similar issue: https://stackoverflow.com/a/4070685

This raises a few questions:
1) Should we implement a mechanism to disable the automatic determination of <release>? E.g. by using the empty string?
  <release></release>
in the pom.

2) Is this a typical situation that we want to avoid? Then maybe determining <release> should be off by default and we could have a extra configuration flag on tycho-compiler-plugin to enable it.
I would say no - because ideally a matching toolchain - in this case for 1.6 - should be used anyways in which case this problem would not occur to begin with.

Patch-set: 1