Update patch set 1
Patch Set 1:

> Looking at the failing tests, I think it's a good thing that it's now failing by default, reminding specific status of this internal package.

Agreed.

> I agree with your proposal. A way to ignore the release flag is still needed though, to have such bundles still able to build. An empty <release> can be good...

I will check what's needed to support this and update the change.

> ..., however, I'm wondering how could users be informed of this being a viable workaround, since we inherit this flag from parent (cannot change the doc) and can't easily catch the case of such internal package being invisible because of <release> being set. Do you have any idea how Tycho could help in that case?

Good question. I've wondered the same, but I don't see a way to detect this situation beforehand. In fact, I don't even see how we could do so after the fact (I'm still wondering why none of these compiler errors made it into any debug log).

So the situation is really hard to deal with usefully.

However: Like I said, If we had a toolchain for 1.6 available, that one would have been used and we would've never noticed this problem.

So maybe the best approach would simply be to not invest too much into this other than mentioning it in the release notes / documentation somewhere.

To get this test running again, we could either
a) Set the proposed <release></release> in the pom.xml. This is not very flexible, because if we ever build with Java 14, this will again fail either way because that concrete class used by the test is gone.
b) Update this test case to something more up-to-date: We could rewrite it to use an internal class that exists in Java 11 and set the BREE to 11. That way we don't have to deal with the 1.6 toolchain situation and no CtSym is involved.

Patch-set: 1