Bug 567760 - Delay 3rd-party p2 artifacts download
Downloads of 3rd party artifacts, so far happening during dependency resolution upon reactor startup, are now delayed to not happen until some piece of code actually requests artifact file location on file-system.
Technically, we propagate a Supplier<File> instead of a File; and introduce some getLocation(boolean fetch) methods to allow invoking this supplier when necessary; and more importantly to not invoke it when
unnecessary.

Change-Id: I6bb62e8cf4f7d4f60440220bd90569b3edf3994b
Signed-off-by: Mickael Istria <mistria@redhat.com>