Bug 567760 - [WIP/DO NOT MERGE] Try delaying download
DO NOT MERGE
Still to do:
* Add tests covering scenario of `mvn validate`
  ** properly validating and reporting issue in case of dep resolution
issue
  ** Not downloading artifacts


Downloads are delayed to not happen until some piece of code actually
requests artifact file location on file-system.
Technically, we propagate a Supplier<File> instead of a File; and
introduce some getLocation(boolean fetch) methods to allow invoking this
supplier when necessary; and more importantly to not invoke it when
unnecessary.

Change-Id: I6bb62e8cf4f7d4f60440220bd90569b3edf3994b
Signed-off-by: Mickael Istria <mistria@redhat.com>