Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I have no clue about what Blackboard-key is for. Did you look through history to find out which change introduces it and get some background about why it's there (and whether it's fine to remove it).
As described before it is currently only used for some kind of mirror action but I haven't found any use beside that single test-util method, as one can see everything, including itests, seems to work without a problem without it.

> Also, please add a bit more content to the commit message: is this change expected to fix the whole issue, or is it just a preliminary change?
The issue itself is referenced in the code:
// TODO 372780 get rid of this special handling of pomDependency artifacts
so it seems related but not actually fixing it.

> And finally, isn't the issue more that the eclipse-repository project dependencies are incorrect in 1st place? I would expect that resolving deps for an eclipse-repository would contain the correct content from target-platform...
I'm not completely sure if I understand the question right, but as you already noticed there is not always an easy answer to this kind of questions :-)

Patch-set: 2