Update patch set 3
Patch Set 3:

As mentioned on another bug, currently tycho does not really support "concurrency" by design but mostly by sync on the mojos so there might be more to consider.

I'm also wondering if tycho must support the concurrency or if not we can delegate this to maven? How is it supposed to work, should plugins/mojo handle concurrent execution by themselves?

For this particular issue, I think the easiest would be to replace WeakHashMap with a ConcurrentHashMap using computeIfAbsent instead of custom checking (no need for guava here IMO).

Patch-set: 3