Update patch set 22
Patch Set 22:

> Patch Set 22:
> 
> If you're in hurry I think you can merge this and we (might) improve later, I won't be able to take a deeper look until tomorrow.

Not that in hurry. It can totally wait for a few days; my ping was more meant to know whether you're interested in reviewing it (in which case we wait for it) or not (in case I'd have merged it). I'm fine waiting if you're interested in reviewing as it brings higher quality early and basically reduce the final amount of work spent in maintenance; but feel free to say no if you're not interested or available in a forseeable future ;)

Patch-set: 22
Reviewer: Gerrit User 587 <587@97ee7c02-f12f-4043-b43e-dea463d88b31>