Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Could you explain please? I can't see how handling two (correlated) field/objects instead of one is more "easily understood".
> The current patch set has 2 simple data structures that are modified in sync with synchronized at 2 places (3 if you count clear). That code is not hard to understand.

That's what I previously commented about those kind of changes: Making think concurrent is more than spreading some synchronized keywords around. Concurrency is all about data structures, if we want concurrent build we have to rethink existing codes in means of finding data structures that behave better in terms concurrent access. If "easy-to-understand" is the primary goal then it would be even more easier to not run the code in parallel at the first place :-)

> Taking just your example of 
> 
> descriptorQueryable can be replaced by 
> (query, monitor) -> query.perform((Iterator<IArtifactDescriptor>) descriptorsMap.entrySet().stream().unordered().flatMap(e->e.getValue().stream()).iterator());
> 
> How does that simplify the code? Also replacing a HashSet::contains with "is there a map value that contains X?", not knowing how that affects performance.

It does remove the need to have two fields kept in sync, just for convenience implementation of two methods that are not proven to be a performance bottleneck as no one has actually profiled there is a need for such optimization here. Both methods are just part of the bloated P2 API that needs to be implemented.

Of course its arguable if stream-way is easier to understand than traditional code we use streams all over the place in tycho so there is no need to worry here IMO.

Patch-set: 1
Reviewer: Gerrit User 171685 <171685@97ee7c02-f12f-4043-b43e-dea463d88b31>