Update patch set 2
Patch Set 2:

Thanks for the feedback. I have thought this was removed with tycho 2.0 so at least we should remove this for 3.0

If everyone is fine with it we can of course try to remove this also in 2.0 I think the first step should be to disable the packing types, then we can remove code step by step.

Patch-set: 2