Update patch set 6
Patch Set 6:

(3 comments)

I'm surprised to see how much change this introduces, I was expecting that we could reuse much more of the current handling of Wrapped artifacts without requiring some new services and so on (I'm not saying this complexity is not necessary, but it was kind of not expected to me). Maybe it's just that I find it a bit tricky to fully understand as a review.
Maybe you could split it into several smaller patches? One for the typo, another for extraction of WrappedArtifact and TargetLocations... That would make this new code easier to understand in review, and later when looking through history.

> @Mickael I tried to add a itest for this but it seems can't figure out how the pom.xml/ .mvn/extension.xml syntax I need to use to properly replace the build versions.
> I tried ${tycho-build-version} and ${tycho-version} but both seem to fail, even though other itests use these without a problem :-\

Sorry I don't know immediately what makes this file content not being filtered properly.
Maybe try to just not use pom-less for this test; it will have the benefit of keeping the test a bit more focused on its goal and may just work more easily.

Patch-set: 6
Reviewer: Gerrit User 587 <587@97ee7c02-f12f-4043-b43e-dea463d88b31>