Update patch set 6
Patch Set 6:

(3 comments)

> Patch Set 6:
> 
> (3 comments)
> 
> I'm surprised to see how much change this introduces, I was expecting that we could reuse much more of the current handling of Wrapped artifacts without requiring some new services and so on (I'm not saying this complexity is not necessary, but it was kind of not expected to me). Maybe it's just that I find it a bit tricky to fully understand as a review.

Yeah I initially has hopped its much less effort. The problem is that the current way works in the wrong direction. There is a (resolved) target that is combined with the dependencies of a maven project.

For this, we have a target (and that code has only limited access to maven API, thus the new service was required) that needs to resolve dependencies of an arbitrary artifact (no project context available) and supply these like regular target content.

That's why I tried to extract as much as possible to share at leas the common parts.

> Maybe you could split it into several smaller patches? One for the typo, another for extraction of WrappedArtifact and TargetLocations... 

I will try this, I just though you might enjoy the whole context :-)

> > @Mickael I tried to add a itest for this but it seems can't figure out how the pom.xml/ .mvn/extension.xml syntax I need to use to properly replace the build versions.
> > I tried ${tycho-build-version} and ${tycho-version} but both seem to fail, even though other itests use these without a problem :-\
> 
> Sorry I don't know immediately what makes this file content not being filtered properly.
> Maybe try to just not use pom-less for this test; it will have the benefit of keeping the test a bit more focused on its goal and may just work more easily.

As we have itests in tycho-extras (I think it would be good to merge them but that's another story) that works that way it must be possible here also. I just can't get whats done differently here, but will give it another try, just wanted to make sure I'm not overseen something obvious.

Patch-set: 6